build: zero non-vendor warnings in default and compute-only flavors
`make` and `make HL_ENABLE_DB=0` previously emitted 11 unique
repo-owned warnings (none of which fired -Werror because the
relevant flags aren't part of -Wall -Wextra -Wpedantic -Wshadow
-Wformat=2). The pollution made real warnings harder to spot in
build logs; now both flavors are silent.

Fixes by category:

* 4× '/*' within block comment (-Wcomment) in include/hull/agent_api.h
  and include/hull/agent_lib.h. The Doxygen prose was writing
  `/_hull/agent/*` and `/_agent/*` to describe URL prefixes, and
  the `/*` got parsed as a nested comment opener. Replaced with
  `/_hull/agent/...` and `/_agent/...`.

* 6× ISO C requires a translation unit to contain at least one
  declaration (-Wempty-translation-unit). The four GPU stubs
  (cap/gpu.c, cap/gpu_wgpu.c, runtime/{lua,js}/mod_gpu.c) become
  empty when HL_ENABLE_GPU=0 (the default); the two agent SQL
  stubs (agent/schema_diff.c, agent/sql.c) become empty when
  HL_ENABLE_DB=0. Each gains a one-line `#else typedef int
  hl_*_disabled; #endif` so the TU is never empty regardless of
  feature flags.

* 1× comparison of integers of different signs (-Wsign-compare)
  in migrate.c:206. `capacity` is `int`; cast to `size_t` to
  match SIZE_MAX/(...).

* 2× format string is not a string literal (-Wformat-nonliteral)
  in serve.c. The format strings come from rxi/log.c's log_Event
  and Keel's KlLogFn callbacks — vsnprintf is doing exactly what
  it's meant to. Wrapped both calls in
  `#pragma GCC diagnostic push/ignored/pop` rather than
  restructure the callback API.

* 1× argument unused during compilation: '-pie'
  (-Wunused-command-line-argument). `-pie` is the macOS default
  since 10.7; passing it explicitly emits a warning on every
  link. Guarded the LDFLAGS addition behind `ifneq
  ($(UNAME_S),Darwin)` so Linux still gets the explicit flag.

35/35 unit suites and 22/22 e2e tests pass in both build flavors.

+++ src/hull/migrate.c
@@ -203,7 +203,7 @@ static int discover_fs_migrations(const char *root_dir, MigrationList *ml)
             continue;
 
         if (ml->count >= capacity) {
-            if (capacity > SIZE_MAX / (2 * sizeof(char *))) {
+            if ((size_t)capacity > SIZE_MAX / (2 * sizeof(char *))) {
                 migration_list_free(ml);
                 closedir(dir);
                 return -1;
